requirements.txt
fastapi
pydantic
uvicorn
pytest
httpx
python-multipart
scikit-learn
numpy
mlflow
joblib
asyncpg
aiokafka
redis
prometheus-client
sentry-sdk[fastapi]
pyjwt
